rename shared preference class
